organise packet capture files
