+++ pyproject.toml
@@ -16,8 +16,6 @@ typer = "0.7.0"
 typer-cli = "^0.0.13"
 sphinx-rtd-theme = "2.0.0"
 
-ovs = "3.1.2"
-
 
 [build-system]
 requires = ["poetry-core"]
